+++ ChangeLog.md
@@ -5,6 +5,13 @@ All notable changes to Hermes will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Added a directory `clients/helpers/` to store some helpers modules that can be used by client plugins.
+- Added `clients.helper.command` helper, to run local commands on client's host.
+
 ## [1.0.0-alpha.3] - 2024-07-12
 
 ### Added